Add files via upload
The task was to transfer the current student database to a new platform created using C++. The new platform is capable of producing an array of students and displaying their information on the console. Additionally, it has features for checking the accuracy of student data, such as email addresses, and can identify any modifications made and notify the user. In essence, this platform has the ability to analyze, structure, and modify student information from lengthy text strings.

This was my submission for C867 at Western Governors University.